code_ ACL_SQ/requirements.txt
openai==1.25.1
pandas==2.2.1
jsonpatch==1.33
jsonpickle==3.3.0
tqdm==4.66.2
transformers==4.45.1
numpy==1.24.4
torch==2.1.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
colorama==0.4.6
langchain-community==0.2.10
evaluate==0.4.1
python-dotenv==1.0.1
langchain-ollama== 0.2.3
sentence-transformers==3.4.1
wheel==0.45.1
wordcloud==1.9.4
tabulate==0.9.0